+++ src/fileio.c
@@ -1182,10 +1182,10 @@ static mame_file *generic_fopen(int pathtype, const char *gamename, const char *
 					if (hash && hash_data_extract_binary_checksum(hash, HASH_CRC, crcs) != 0)
 					{
 						/* Store the CRC in a single DWORD */
-						crc = ((unsigned long)crcs[0] << 24) |
-							  ((unsigned long)crcs[1] << 16) |
-							  ((unsigned long)crcs[2] <<  8) |
-							  ((unsigned long)crcs[3] <<  0);
+						crc = ((UINT32)crcs[0] << 24) |
+							  ((UINT32)crcs[1] << 16) |
+							  ((UINT32)crcs[2] <<  8) |
+							  ((UINT32)crcs[3] <<  0);
 					}
 
 					hash_data_clear(file.hash);
