+++ src/machine/stvcd.c
@@ -65,14 +65,14 @@ CD: End data transfer (PC=4232)
 static cdrom_file *cdrom = (cdrom_file *)NULL;
 
 static void cd_readTOC(void);
-static void cd_readblock(unsigned long fad, unsigned char *dat);
+static void cd_readblock(UINT32 fad, unsigned char *dat);
 static void cd_playdata(void);
 
 typedef struct
 {
 	unsigned char flags;		// iso9660 flags
-	unsigned long length;		// length of file
-	unsigned long firstfad;		// first sector of file
+	UINT32 length;		// length of file
+	UINT32 firstfad;		// first sector of file
 	unsigned char name[128];
 } direntryT;
 
@@ -114,8 +114,8 @@ static UINT32 dbufptr;
 static UINT32 dbuflen;
 
 // iso9660 utilities
-static void read_new_dir(unsigned long fileno);
-static void make_dir_current(unsigned long fad);
+static void read_new_dir(UINT32 fileno);
+static void make_dir_current(UINT32 fad);
 
 static direntryT curroot;		// root entry of current filesystem
 static direntryT *curdir;		// current directory
@@ -306,7 +306,7 @@ static UINT16 cd_readWord(UINT32 addr)
 
 static UINT32 cd_readLong(UINT32 addr)
 {
-	unsigned long rv;
+	UINT32 rv;
 
 	switch (addr & 0xffff)
 	{
@@ -336,7 +336,7 @@ static UINT32 cd_readLong(UINT32 addr)
 
 static void cd_writeWord(UINT32 addr, UINT16 data)
 {
-	unsigned long temp;
+	UINT32 temp;
 
 	switch(addr & 0xffff)
 	{
@@ -906,10 +906,10 @@ WRITE32_HANDLER( stvcd_w )
 }
 
 // iso9660 parsing
-static void read_new_dir(unsigned long fileno)
+static void read_new_dir(UINT32 fileno)
 {
 	int foundpd, i;
-	unsigned long cfad, dirfad;
+	UINT32 cfad, dirfad;
 	unsigned char sect[2048];
 
 	if (fileno == 0xffffff)
@@ -966,7 +966,7 @@ static void read_new_dir(unsigned long fileno)
 			// easy to fix, but make sure we *need* to first
 			if (curroot.length > 14336)
 			{
-				fatalerror("ERROR: root directory too big (%ld)", curroot.length);
+				fatalerror("ERROR: root directory too big (%ld)", (unsigned long)curroot.length);
 			}
 
 			// done with all that, read the root directory now
@@ -977,17 +977,17 @@ static void read_new_dir(unsigned long fileno)
 	{
 		if (curdir[fileno].length > 14336)
 		{
-			fatalerror("ERROR: new directory too big (%ld)!", curdir[fileno].length);
+			fatalerror("ERROR: new directory too big (%ld)!", (unsigned long)curdir[fileno].length);
 		}
 		make_dir_current(curdir[fileno].firstfad);
 	}
 }
 
 // makes the directory pointed to by FAD current
-static void make_dir_current(unsigned long fad)
+static void make_dir_current(UINT32 fad)
 {
 	int i;
-	unsigned long nextent, numentries;
+	UINT32 nextent, numentries;
 	unsigned char sect[14336];
 	direntryT *curentry;
 
@@ -1161,7 +1161,7 @@ static void cd_playdata(void)
 }
 
 // loads a single sector off the CD, anywhere from FAD 150 on up
-static void cd_readblock(unsigned long fad, unsigned char *dat)
+static void cd_readblock(UINT32 fad, unsigned char *dat)
 {
 	if (cdrom)
 	{
